Update examples.tscn, shield.gd, shield_base.gdshaderinc, and 1 more file
Shield.gd script now adds a new ripple effect when the object that created the ripple is still touching the shield.

Created a start for double checking that all objects touching the shield have a ripple effect.

Later when creating a way for ripple to change characteristics based on the size and speed of the object that hit the shield, also have factor variable on line 298 change as well so the secondary ripples flow smoothly with the first ripple.