Update examples.tscn and shield.gd
More work, more stuff works, and more bugs to find. Something is making some of the ripples look not as smooth as they should, and in the example scene the big square is acting like their is a hug wave in it and the little example the waves stop for 1 or 2 seconds.